moving files for oc and manifolds
